+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Added `content.get_relations` and `content.get_back_relations` to easily
   get relations and back relations on an object.
   [gbastien]
+- Do not break in `xhtml.storeImagesLocally` if image URL
+  contains non-ASCII characters.
+  [gbastien]
 
 0.32 (2020-09-10)
 -----------------